- Moved ArrayIterator to its own file in utils. - Added validation for the 'size'.
